\name{RcmdrPlugin.TeachingDemos-package}
\alias{RcmdrPlugin.TeachingDemos-package}
\alias{RcmdrPlugin.TeachingDemos}
\docType{package}
\title{
Install the Demos Rcmdr Plug-In
}
\description{
Provides an Rcmdr "plug-in" based on the TeachingDemos package, and is primarily for illustrative purposes.
}
\details{
\tabular{ll}{
Package: \tab RcmdrPlugin.TeachingDemos\cr
Type: \tab Package\cr
Version: \tab 1.2-0\cr
Date: \tab 2020-12-02\cr
License: \tab GPL version 2 or newer\cr
}
}

\author{
John Fox <jfox@mcmaster.ca>

Maintainer: John Fox <jfox@mcmaster.ca>
}

\keyword{ package }
\seealso{
\code{\link[Rcmdr:Rcmdr-package]{Rcmdr}}.
}

\examples{
if (interactive()){ # load the Rcmdr and the plug-in
    library(RcmdrPlugin.TeachingDemos)
    }
}
